// application/types/ParsedItemType.ts
export type ParsedItemType = {
  representativeThumbnailNoAuth: string
  title: string
  itemId: string
  durationSeconds: string
  originalWidth: string
  originalHeight: string
  created: string
  mimeType: string
  originalVideoCodec: string
  startTimeCode: string
  size: number
  item_videoId: string
  item_airDate: string
  item_slug: string
}